requirements.txt
flask
google-cloud-aiplatform
python-dotenv
